Updated reporesentation for TTP data and EA interfaces.

+++ EA/DataTTP/Node.cs
@@ -15,7 +15,7 @@ namespace EA.DataTTP
             this.Y = 0;
         }
 
-        public Node(int index, decimal x, decimal y)
+        public Node(int index, double x, double y)
         {
             Index = index;
             X = x;
@@ -23,7 +23,7 @@ namespace EA.DataTTP
         }
 
         public int Index { get; set; }
-        public decimal X { get; set; }
-        public decimal Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
     }
 }
